V0.1.3 Update
Have updated system to now have a start date for tasks so when showing
the previous percentage it does not decrease as more tasks are added.

Also added a new field called "help" ready to add help text to tasks
allowing you to describe what is needing to be done on each task.